Add Windows test project

+++ src/cauchy_256.cpp
@@ -365,7 +365,7 @@ static void GFC256Init()
 	// For each other y value,
 	for (int y = 1; y < 256; ++y) {
 		// Calculate log(y) for mult and 255 - log(y) for div
-		const u8 log_y = GFC256_LOG_TABLE[y];
+		const u8 log_y = (u8)GFC256_LOG_TABLE[y];
 		const u8 log_yn = 255 - log_y;
 
 		// Next subtable
